Moved DocActivityType, DocStatus & MethodBadge into global components folder
- Updated tests
- Updated corresponding index.ts to reflect changes
