+++ packages/api-explorer/src/components/DocRateLimited/DocRateLimited.spec.tsx
@@ -1,0 +1,50 @@
+/*
+
+ MIT License
+
+ Copyright (c) 2020 Looker Data Sciences, Inc.
+
+ Permission is hereby granted, free of charge, to any person obtaining a copy
+ of this software and associated documentation files (the "Software"), to deal
+ in the Software without restriction, including without limitation the rights
+ to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ copies of the Software, and to permit persons to whom the Software is
+ furnished to do so, subject to the following conditions:
+
+ The above copyright notice and this permission notice shall be included in all
+ copies or substantial portions of the Software.
+
+ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ SOFTWARE.
+
+ */
+import React from 'react'
+import { renderWithTheme } from '@looker/components-test-utils'
+import { screen, fireEvent, waitFor } from '@testing-library/react'
+
+import { api } from '../../test-data'
+import { DocRateLimited } from './DocRateLimited'
+
+describe('DocRateLimited', () => {
+  const rateLimitedMethod = api.methods.test_connection
+  test('it renders', () => {
+    renderWithTheme(<DocRateLimited method={rateLimitedMethod} />)
+    expect(screen.getByText('rate_limited')).toBeInTheDocument()
+  })
+
+  test('it displays a tooltip with the right information on hover', async () => {
+    renderWithTheme(<DocRateLimited method={rateLimitedMethod} />)
+    const badge = screen.getByText('rate_limited')
+    fireEvent.mouseOver(badge)
+    await waitFor(() => {
+      expect(screen.getByRole('tooltip')).toHaveTextContent(
+        'This endpoint is rate limited.'
+      )
+    })
+  })
+})
